+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Added support for resolving array of names
+* Look for namespace short name as default resolution
+
 ## v0.2.16 (2023-10-30)
 * Improved default interface resolution